Fix project updating from the notification
