+++ package.json
@@ -1213,6 +1213,7 @@
     "@types/mocha": "^5.2.5",
     "@types/node": "^8.10.51",
     "@types/semver": "^7.3.8",
+    "@types/sinon": "^10.0.12",
     "@types/vscode": "^1.65.0",
     "@types/winreg": "^1.2.30",
     "@types/winston": "^2.4.4",
@@ -1225,6 +1226,7 @@
     "minimist": ">=1.2.6",
     "mocha": "^9.2.0",
     "request": "^2.88.2",
+    "sinon": "^14.0.0",
     "ts-loader": "^9.2.6",
     "tslint": "^5.11.0",
     "typescript": "^4.2.4",
